Fixed bug in type inference.
Although the type inference takes into account the type signature of
functions during inference, the resulting declarations got the inferred
type of the function rule, which could be more general.

The new implementation now checks whether the type provided by the
function signature is an instance of the inferred type and keeps it.