Clean up explicite type annotations involving `length`
In an earlier version of Prelude.curry, `length` was missing an explicite type annotations, resulting in the following type:

length :: Num a => [b] -> a

In combination with numerical literals, type class methods like `(<)` could not infer the needed dictionary leading to an "ambiguous contexts" type error. In that state of the type classes branch, defaulting was not yet implemented, thus explicite type annotations were added in order to compile kics2 accordingly.